requirements.txt
PyAutoGUI.egg==info
selenium==4.0.0a1
